Major rewrite of EWF code.
Most of the EWF (Ecasound Wave File) support has been rewritten.
This rewrite was done to increase code robustness and readability.
Many knowns bugs should be now fixed, and hopefully not that many
new bugs were introduced.

At the same time, the ability to write to EWF files has been
deprecated as a feature and will be removed in a later release.