Fixed mp3 encoding/output with lame 3.98 and newer
The command-line syntax of 'lame' mp3 encoder, and semantics
for old options, have changed slightly in lame 3.97 and 3.98.
This patch updates the default command used to launch lame, and
also provides some extra code to support old version of lame,
as well as to support old ecasoundrc files (to avoid problems
when upgrading Ecasound).