Updates to kvu_dbc.h implementation
The formatting of warnings, emitted when design-by-contract
assertions fail, has been slightly updated and the pretty
printing is now implemented separately in kvu_dbc.cpp (versus
in the header as before).

As an additional feature, if ecasound is compiled with
debug features enabled ('--enable-debug' to configure),
a function stack backtrace is printed out whenever a DBC
assertion fails. Currently this feature is only available
with glibc (2.1 and newer).