New async interface for ECA_ENGINE and edit objects for ECA_CHAINSETUP
Replace the old async message queue interface of ECA_ENGINE with
a more generic one (implemented on top of kvu_message_queue.h). This
patch also updates all the client code to use the new engine interface.

As a directly related change, eca-chainsetup-edit.h introduces
a type for describing chainsetup edits as objects that can
be passed through message queues. This allows to reuse the same
code both on the real-time path from ECA_ENGINE, as well as
for non-real-time editing from ECA_CONTROL. Additionally this
allows to extend the stateless interface (not depending on
the current set of selected objects) of ECA_CONTROL.