eca-chainsetup: Adjust checks for rt capabilities
Relax the checks for real-time capabilities when
deciding on which buffering mode to select for the given
chainsetup. Specifically do not require support for
memory locking in order to enable "rt" and "rtlowlatency"
buffering modes.

Also a warning message is now printed, if the chainsetup
would benefit from real-time capabilities (scheduling
and memory locking), but rtcaps are not allowed or
otherwise not available to the current process.