eca-chainsetup: Fix bug with removing audio objects
When removing non-realtime audio objects, with disk
doublebuffering in use, the attached proxy object (of
type AUDIO_IO_DB_CLIENT*) was not deleted at all.

Proper cleanup was still done when removing the whole
chainsetup, but programs doing a lot of object removals
risk leaking memory (until the chainsetup was deleted),
and may have hit a segfault in some scenarios.

Ecasignalview is one application that was affected by
this bug.