eca-sample-conversion: make int-float conversions symmetric
Asymmetricity of the conversion routines (using different scaling
for positive and negative values), can cause distortion to
the signal. This patch modifies all the conversion routines to
be asymmetric.

To handle float-to-int conversion in the case of 0dBFS signals
created in floating point domain, introduce clipping for positive
samples near 0dBFS (with different threshold based on target
int precision).