audioio: update locked_audio_format() documentation
Update the documentation to match the current implementation.
Many objects return true for locked_audio_format() despite
allowing to set some (but not all!) of the audio parameters.

A common case is fixing the sample type, but allowing to freely
set sampling rate and channel count. Various other combinations
are used in the codebase.