eca-audio-format: Add a FIXME related to Sample enums
The Sample_foo enums should be moved to sample-specs.h,
so that that codebase wide defaults (e.g. use of float or
double) can be expressed in terms of these enums. This
would allow other places in the code to use express that
they wish to use the same sample type as is used by the engine
(without hardcoding any values).