build: start maintaining release notes in RELNOTES file
Maintain release notes for the current version in development
in the RELNOTES files. When the release is done, this file is
sent as the announcement mail.

To reduce duplicated work in tracking changes, NEWS file scope
is reduce to just listing releases and release dates. Nowadays
the most detailed change history is anyways conveniently
available in the form of git commit history.