ecasound: fix a race with sending SIGHUP to signal watchdog
Fixes a race in sending SIGHUP to the signal watchdog thread, and
unblocking signals just before terminating the process.

On some systems, unblocking of SIGHUP could result in delivering
the signal to the wrong thread (not to the watchdog thread as
was expected by the code), and this would leave the process hang,
with main thread waiting for signal watchdog (in pthread_join())
and signal watchdog blocked on sigwait().

Bug reported by Alessandro Ghedini.