Add draining support to engine for real-time audio objects
When processing is stopped after running out of data from all inputs,
or reaching the set fixed processing time, make sure all already buffered
audio is played out to real-time devices (e.g. to audio hardware) before
the devices are closed.

In case of errors, or performing seeks, this mechanism is not used as
it could cause unnecessary delays.

Both the default and the JACK driver backends are updated.