Added 'stop-sync' command
A blocking variant of the existing 'stop' command. This is
useful in ECI apps that need to make sure engine has really
stopped before they start modifying the chainsetup contents.

The standard 'stop' command is asynchronous, so it is just
a request to stop processing, but upon return to the ECI call,
the engine may still run for a while.