libecasound: add on/off param to c-bypass/c-mute
Align c-bypass and c-mute commands with the new cop-bypass
command, allowing to specify "on" or "off". If no param
is given, behaviour defaults to the old toggle the previous
value ("toggle" param can also be given explicitly).

Under the hood, c-bypass and c-mute are now performed
as eca-chainsetup-edit objects, simplifying the implementation
and getting rid of unnecessary code from ECA_ENGINE.