kvutils: added note about pitfalls in kvu_pthread_timed_wait()
To prevent misuse of kvu_pthread_timed_wait() function, add
a note about the limitation. Notably the condition cannot
be checked with mutex held, so going to sleep is not race
free with this function. It still has its uses (when losing
a condition signal is ok), but not as a generic wrapper to
pthread_cond_timedwait().