libecasound: refactor stopping-for-edit logic
This improves on the solution added in commit:

408f38ee82216601553f3fee01aab96be9314011
'Add is_started() to eca-engine and eca-control'

The problem with the earlier implementation was that even though it
guaranteed that the engine was stopped, it could not guarantee
that the engine would not immediately be restarted (while edit was
still in progress). And as stops can be caused by many different
events (for example: external JACK transport state change), blocking
these requires a more robust design.

This patch uses the existing 'ep_edit_lock' and extends
the interfaces so control client can wait for the edit lock,
in a race-free manner. Edit locks are requested by ECA_CONTROL
(one per ECA_ENGINE) only, so effectively this allows to
implement a reliable synchronization point.

To make modification, ECA_CONTROL will send a sequence of 'stop' and
'edit_lock' commands and wait on the edit lock signal. Once engine
reaches the sync point, ECA_CONTROL knows that
  a) engine is stopped,
  b) all prior commands have been processed, and
  c) nobody can start it.

ECA_CONTROL can now safely modify the setup, before sending
the 'edit_unlock' and 'start' restart sequence.