libecasound: lower debug level for SCHED_FIFO usage msg
Do not print an info level "Using realtime-scheduling (SCHED_FIFO)"
message every time the engine is prepared. This gets fairly
annoying when used with for example JACK's transport control,
or doing a lot of seeks, or live edits to the chainsetup.
Without the patche, the message was printed to console on each
restart of the engine.