osc-gen: fix seeking bugs, refactor implementation
Major rework to the implementation of '-kog' and
'-kf' controller objects. The main objective was to add
proper support for seeking (so that these controllers behave
deterministically with seeks), but in the process, some bugs
were fixed and in general, the performance is now somewhat
better (albeit still far from optimal).

Closes sf.net bugs:
  - sfbug:2013828, https://sourceforge.net/tracker/?func=detail&aid=2013828&group_id=4605&atid=104605
  - sfbug:2013824, https://sourceforge.net/tracker/?func=detail&aid=2013824&group_id=4605&atid=104605