sndfile: fix seeking beyond end-of-file
For input files handled with libsndfile, after a seek attempt beyond
end-of-file, set the current position of the file to to last sample+1
(i.e. to end-of-file).

Without this fix, seeks beyond end-of-file failed (and failure was
reported via a message), but position remained where it was before the
seek attempt. This was highly confusing when used with multiple inputs
of different lengths (a global seek that was beyond end-of-file for one
input, would cause the set of files to be played out of sync).