rename is_engine_running() to is_engine_ready_for_commands()
The old ECA_CONTROL::is_engine_running() was super-confusingly
named as this had nothing to do whether engine state was "running"
or not. Insted this boolean check refers to the engine thread
and whether it is running and can be reached via sending messages.
Using the same "running" word was a bad chose, so now renaming
the ECA_CONTROL method accordingly.