refactor cop-add/ctrl-add implementation for connected chainsetups
Third attempt at refactoring how cop-add/ctrl-add are executed when
target is a connected chainsetup. Previous attempt was last month with
commit ca925a389154774ca7ff69d2f141f1c9000a2050 but that has proven to
be a failed approach.

With this patch, we get rid of the "stop engine, wait for engine, do
edits in the control thread, restart engine if it was previously
running" logic altogether, including engine edit locks (ep_edit_lock)
and ECA_CONTROL::cond_stop_for_editing().

Instead, cop-add/ctrl-add can be now done via edit objects
(eca-chainsetup-edit.h), and executed completely in the engine thread
context. The edit is now an atomic operation from engine point of view,
so there's no need to synchronize state across threads, which has been
the biggest cause for regression in the past in this area.

Downside: cop-remove and ctrl-remove are no longer allowed on
a disconnected chainsetup. Support for this can be added later if
the general approach proves solid.