libecasound: fix ai/ao-remove bug with duplicate labels
The ECA_CHAINSETUP interface to remove audio inputs and outputs caused
bugs when there were multiple objects with the same label. As a result,
wrong object was removed.

This patch fixes by changing the interface to use object pointers
instead of string labels.

Bug reported by Rocco on ecasound-list on 2013/04.