build: do C++ builds in -std=c++98 mode by default
Codebase builds ok with -std=c++14, but it does not use any
features of C++11 or newer yet, and it does emit warnings in
C++11 mode (e.g. use of dynamic exception throw specifications).

Codebase will be migrated to C++11 later, but before that,
the obvious warnings at least need to be removed.