eca-chainsetup: remove inaccurate dbc warning in device open
A false design-by-contract (DBC) warnings was emitted when
opening a file or audio device that modified audio parameters
to closest matching value, but did not report
'locked_audio_format()==true'. This check is not valid as
modification of audio parameters is allowed even for objects
that do not have a locked audio format.