pyecasound: remove C pyecasound implementation
The C implementation of ECI API using Python module interface
was deprecated in 2012 in 2.9.0. It has still been possible
to enable it with "--enable-python=iwantc".

With upgrade of Ecasound's Python support to Python 3.x,
the C pyecasound implementation would require further
updates to keep using it with the new Python3. Instead,
the C implementation is now dropped completely.
The pyeca.py implementation is mature and supports both
Python2.7 and Python3.