libecasound: LADSPA: fix bug in parsing output capability
Fix an old bug in parsing LADSPA plugin hints on port
properties. The check whether a LADSPA port is an output or
not, was incorrect and was always set as false.

This bug did not prevent using LADSPA plugins with output
ports, but the ecasound description of the plugin (as e.g.
in output of "map-ladspa-list") was wrong.

Bug was found based on warning from GCC 8.3.0.