libecasound: do not normalize float file outputs to [-1,1]
Ecasound has traditionally normalized sample data before
export to [-1,1] value range and not allowed samples above 0dBFS.

There's really no good reason for doing this, especially as
ecasound imports float audio files with values oer 0dBFS just
fine, so export should behave similarly.

This is a minimal patch to change the behaviour. There is a lot
of old cruft in samplebuffer.cpp that should be cleaned up, but
I resisted the temptation and kept this patch simple as possible.